Simplify property updates
