+++ src/propertyTypes.ts
@@ -1,0 +1,9 @@
+/** THIS FILE IS AUTOMATICALLY GENERATED. See tools/parseDefaultProperties.ts for more information */
+export type BooleanProperty = "addChatCommandLine" | "addCreationQueue" | "addStatusBarToFrames" | "allowCloseableDesktopTabs" | "allowNegativeTally" | "allowNonMoodBurning" | "allowSummonBurning" | "allowSocketTimeout" | "autoHighlightOnFocus" | "broadcastEvents" | "cacheMallSearches" | "chatBeep" | "chatLinksUseRelay" | "cloverProtectActive" | "compactChessboard" | "connectViaAddress" | "copyAsHTML" | "customizedTabs" | "debugBuy" | "debugConsequences" | "debugFoxtrotRemoval" | "debugPathnames" | "gapProtection" | "greenScreenProtection" | "guiUsesOneWindow" | "hideServerDebugText" | "logAcquiredItems" | "logBattleAction" | "logBrowserInteractions" | "logChatMessages" | "logChatRequests" | "logCleanedHTML" | "logDecoratedResponses" | "logFamiliarActions" | "logGainMessages" | "logReadableHTML" | "logPreferenceChange" | "logMonsterHealth" | "logReverseOrder" | "logStatGains" | "logStatusEffects" | "logStatusOnLogin" | "macroDebug" | "macroLens" | "mementoListActive" | "mergeHobopolisChat" | "printStackOnAbort" | "protectAgainstOverdrink" | "proxySet" | "relayAddSounds" | "relayAddsCustomCombat" | "relayAddsDiscoHelper" | "relayAddsGraphicalCLI" | "relayAddsQuickScripts" | "relayAddsRestoreLinks" | "relayAddsUpArrowLinks" | "relayAddsUseLinks" | "relayAddsWikiLinks" | "relayAllowRemoteAccess" | "relayBrowserOnly" | "relayFormatsChatText" | "relayHidesJunkMallItems" | "relayMaintainsEffects" | "relayMaintainsHealth" | "relayMaintainsMana" | "relayOverridesImages" | "relayRunsAfterAdventureScript" | "relayRunsBeforeBattleScript" | "relayRunsBeforePVPScript" | "relayScriptButtonFirst" | "relayTextualizesEffects" | "relayTrimsZapList" | "relayUsesInlineLinks" | "relayUsesIntegratedChat" | "relayWarnOnRecoverFailure" | "removeMalignantEffects" | "saveSettingsOnSet" | "sharePriceData" | "showAllRequests" | "showAnnouncements" | "showExceptionalRequests" | "stealthLogin" | "svnInstallDependencies" | "svnShowCommitMessages" | "svnUpdateOnLogin" | "switchEquipmentForBuffs" | "syncAfterSvnUpdate" | "useChatToolbar" | "useContactsFrame" | "useDevProxyServer" | "useDockIconBadge" | "useHugglerChannel" | "useImageCache" | "useLastUserAgent" | "useNaiveSecureLogin" | "useShinyTabbedChat" | "useSystemTrayIcon" | "useTabbedChatFrame" | "useToolbars" | "useZoneComboBox" | "verboseSpeakeasy" | "verboseFloundry" | "_announcementShown" | "_svnRepoFileFetched" | "_svnUpdated" | "antagonisticSnowmanKitAvailable" | "arcadeGameHints" | "armoryUnlocked" | "autoCraft" | "autoQuest" | "autoEntangle" | "autoGarish" | "autoManaRestore" | "autoFillMayoMinder" | "autoPinkyRing" | "autoPlantHardcore" | "autoPlantSoftcore" | "autoPotionID" | "autoRepairBoxServants" | "autoSatisfyWithCloset" | "autoSatisfyWithCoinmasters" | "autoSatisfyWithMall" | "autoSatisfyWithNPCs" | "autoSatisfyWithStash" | "autoSatisfyWithStorage" | "autoSetConditions" | "autoSphereID" | "autoSteal" | "autoTuxedo" | "backupCameraReverserEnabled" | "badMoonEncounter01" | "badMoonEncounter02" | "badMoonEncounter03" | "badMoonEncounter04" | "badMoonEncounter05" | "badMoonEncounter06" | "badMoonEncounter07" | "badMoonEncounter08" | "badMoonEncounter09" | "badMoonEncounter10" | "badMoonEncounter11" | "badMoonEncounter12" | "badMoonEncounter13" | "badMoonEncounter14" | "badMoonEncounter15" | "badMoonEncounter16" | "badMoonEncounter17" | "badMoonEncounter18" | "badMoonEncounter19" | "badMoonEncounter20" | "badMoonEncounter21" | "badMoonEncounter22" | "badMoonEncounter23" | "badMoonEncounter24" | "badMoonEncounter25" | "badMoonEncounter26" | "badMoonEncounter27" | "badMoonEncounter28" | "badMoonEncounter29" | "badMoonEncounter30" | "badMoonEncounter31" | "badMoonEncounter32" | "badMoonEncounter33" | "badMoonEncounter34" | "badMoonEncounter35" | "badMoonEncounter36" | "badMoonEncounter37" | "badMoonEncounter38" | "badMoonEncounter39" | "badMoonEncounter40" | "badMoonEncounter41" | "badMoonEncounter42" | "badMoonEncounter43" | "badMoonEncounter44" | "badMoonEncounter45" | "badMoonEncounter46" | "badMoonEncounter47" | "badMoonEncounter48" | "barrelShrineUnlocked" | "bigBrotherRescued" | "blackBartsBootyAvailable" | "bondAdv" | "bondBeach" | "bondBeat" | "bondBooze" | "bondBridge" | "bondDesert" | "bondDR" | "bondDrunk1" | "bondDrunk2" | "bondHoney" | "bondHP" | "bondInit" | "bondItem1" | "bondItem2" | "bondItem3" | "bondJetpack" | "bondMartiniDelivery" | "bondMartiniPlus" | "bondMartiniTurn" | "bondMeat" | "bondMox1" | "bondMox2" | "bondMPregen" | "bondMus1" | "bondMus2" | "bondMys1" | "bondMys2" | "bondSpleen" | "bondStat" | "bondStat2" | "bondStealth" | "bondStealth2" | "bondSymbols" | "bondWar" | "bondWeapon2" | "bondWpn" | "booPeakLit" | "bootsCharged" | "breakfastCompleted" | "burrowgrubHiveUsed" | "canteenUnlocked" | "chaosButterflyThrown" | "chatbotScriptExecuted" | "chateauAvailable" | "chatLiterate" | "chatServesUpdates" | "checkJackassHardcore" | "checkJackassSoftcore" | "clanAttacksEnabled" | "coldAirportAlways" | "considerShadowNoodles" | "controlRoomUnlock" | "concertVisited" | "controlPanel1" | "controlPanel2" | "controlPanel3" | "controlPanel4" | "controlPanel5" | "controlPanel6" | "controlPanel7" | "controlPanel8" | "controlPanel9" | "corralUnlocked" | "dailyDungeonDone" | "dampOldBootPurchased" | "daycareOpen" | "demonSummoned" | "dinseyAudienceEngagement" | "dinseyGarbagePirate" | "dinseyRapidPassEnabled" | "dinseyRollercoasterNext" | "dinseySafetyProtocolsLoose" | "doghouseBoarded" | "dontStopForCounters" | "drippingHallUnlocked" | "drippyShieldUnlocked" | "edUsedLash" | "eldritchFissureAvailable" | "eldritchHorrorAvailable" | "essenceOfAnnoyanceAvailable" | "essenceOfBearAvailable" | "expressCardUsed" | "falloutShelterChronoUsed" | "falloutShelterCoolingTankUsed" | "fistTeachingsHaikuDungeon" | "fistTeachingsPokerRoom" | "fistTeachingsBarroomBrawl" | "fistTeachingsConservatory" | "fistTeachingsBatHole" | "fistTeachingsFunHouse" | "fistTeachingsMenagerie" | "fistTeachingsSlums" | "fistTeachingsFratHouse" | "fistTeachingsRoad" | "fistTeachingsNinjaSnowmen" | "flickeringPixel1" | "flickeringPixel2" | "flickeringPixel3" | "flickeringPixel4" | "flickeringPixel5" | "flickeringPixel6" | "flickeringPixel7" | "flickeringPixel8" | "frAlways" | "frCemetaryUnlocked" | "friarsBlessingReceived" | "frMountainsUnlocked" | "frSwampUnlocked" | "frVillageUnlocked" | "frWoodUnlocked" | "getawayCampsiteUnlocked" | "ghostPencil1" | "ghostPencil2" | "ghostPencil3" | "ghostPencil4" | "ghostPencil5" | "ghostPencil6" | "ghostPencil7" | "ghostPencil8" | "ghostPencil9" | "gingerAdvanceClockUnlocked" | "gingerBlackmailAccomplished" | "gingerbreadCityAvailable" | "gingerExtraAdventures" | "gingerNegativesDropped" | "gingerSewersUnlocked" | "gingerSubwayLineUnlocked" | "gingerRetailUnlocked" | "glitchItemAvailable" | "grabCloversHardcore" | "grabCloversSoftcore" | "guideToSafariAvailable" | "guyMadeOfBeesDefeated" | "hardcorePVPWarning" | "harvestBatteriesHardcore" | "harvestBatteriesSoftcore" | "hasBartender" | "hasChef" | "hasCocktailKit" | "hasDetectiveSchool" | "hasOven" | "hasRange" | "hasShaker" | "hasSushiMat" | "haveBoxingDaydreamHardcore" | "haveBoxingDaydreamSoftcore" | "hermitHax0red" | "holidayHalsBookAvailable" | "horseryAvailable" | "hotAirportAlways" | "implementGlitchItem" | "itemBoughtPerAscension637" | "itemBoughtPerAscension8266" | "itemBoughtPerAscension10790" | "itemBoughtPerAscension10794" | "itemBoughtPerAscension10795" | "itemBoughtPerCharacter6423" | "itemBoughtPerCharacter6428" | "itemBoughtPerCharacter6429" | "kingLiberated" | "lastPirateInsult1" | "lastPirateInsult2" | "lastPirateInsult3" | "lastPirateInsult4" | "lastPirateInsult5" | "lastPirateInsult6" | "lastPirateInsult7" | "lastPirateInsult8" | "lawOfAveragesAvailable" | "leafletCompleted" | "libraryCardUsed" | "lockPicked" | "loginRecoveryHardcore" | "loginRecoverySoftcore" | "lovebugsUnlocked" | "loveTunnelAvailable" | "lowerChamberUnlock" | "makePocketWishesHardcore" | "makePocketWishesSoftcore" | "manualOfNumberologyAvailable" | "mappingMonsters" | "mapToAnemoneMinePurchased" | "mapToKokomoAvailable" | "mapToMadnessReefPurchased" | "mapToTheDiveBarPurchased" | "mapToTheMarinaraTrenchPurchased" | "mapToTheSkateParkPurchased" | "maraisBeaverUnlock" | "maraisCorpseUnlock" | "maraisDarkUnlock" | "maraisVillageUnlock" | "maraisWildlifeUnlock" | "maraisWizardUnlock" | "maximizerAlwaysCurrent" | "maximizerCreateOnHand" | "maximizerCurrentMallPrices" | "maximizerFoldables" | "maximizerIncludeAll" | "maximizerNoAdventures" | "middleChamberUnlock" | "moonTuned" | "neverendingPartyAlways" | "odeBuffbotCheck" | "oilPeakLit" | "oscusSodaUsed" | "outrageousSombreroUsed" | "pathedSummonsHardcore" | "pathedSummonsSoftcore" | "popularTartUnlocked" | "prAlways" | "prayedForGlamour" | "prayedForProtection" | "prayedForVigor" | "pyramidBombUsed" | "ROMOfOptimalityAvailable" | "rageGlandVented" | "readManualHardcore" | "readManualSoftcore" | "relayShowSpoilers" | "relayShowWarnings" | "rememberDesktopSize" | "restUsingChateau" | "restUsingCampAwayTent" | "requireBoxServants" | "requireSewerTestItems" | "safePickpocket" | "schoolOfHardKnocksDiplomaAvailable" | "serverAddsCustomCombat" | "SHAWARMAInitiativeUnlocked" | "showGainsPerUnit" | "showIgnoringStorePrices" | "showNoSummonOnly" | "showTurnFreeOnly" | "sleazeAirportAlways" | "snojoAvailable" | "sortByRoom" | "spacegateAlways" | "spacegateVaccine1" | "spacegateVaccine2" | "spacegateVaccine3" | "spaceInvaderDefeated" | "spelunkyHints" | "spiceMelangeUsed" | "spookyAirportAlways" | "stenchAirportAlways" | "stopForFixedWanderer" | "styxPixieVisited" | "suppressInappropriateNags" | "suppressPotentialMalware" | "suppressPowerPixellation" | "telegraphOfficeAvailable" | "telescopeLookedHigh" | "timeTowerAvailable" | "trackLightsOut" | "uneffectWithHotTub" | "universalSeasoningActive" | "universalSeasoningAvailable" | "useCrimboToysHardcore" | "useCrimboToysSoftcore" | "verboseMaximizer" | "visitLoungeHardcore" | "visitLoungeSoftcore" | "visitRumpusHardcore" | "visitRumpusSoftcore" | "voteAlways" | "wildfireBarrelCaulked" | "wildfireDusted" | "wildfireFracked" | "wildfirePumpGreased" | "wildfireSprinkled" | "yearbookCameraPending" | "_affirmationCookieEaten" | "_affirmationHateUsed" | "_akgyxothUsed" | "_alienAnimalMilkUsed" | "_alienPlantPodUsed" | "_allYearSucker" | "_aprilShower" | "_armyToddlerCast" | "_baconMachineUsed" | "_bagOfCandy" | "_bagOfCandyUsed" | "_bagOTricksUsed" | "_ballastTurtleUsed" | "_ballInACupUsed" | "_ballpit" | "_barrelPrayer" | "_beachCombing" | "_bendHellUsed" | "_blankoutUsed" | "_bonersSummoned" | "_borrowedTimeUsed" | "_bowleggedSwaggerUsed" | "_bowlFullOfJellyUsed" | "_boxOfHammersUsed" | "_brainPreservationFluidUsed" | "_brassDreadFlaskUsed" | "_cameraUsed" | "_canSeekBirds" | "_carboLoaded" | "_cargoPocketEmptied" | "_ceciHatUsed" | "_chateauDeskHarvested" | "_chateauMonsterFought" | "_chronerCrossUsed" | "_chronerTriggerUsed" | "_chubbyAndPlumpUsed" | "_circleDrumUsed" | "_clanFortuneBuffUsed" | "_claraBellUsed" | "_coalPaperweightUsed" | "_cocoaDispenserUsed" | "_cocktailShakerUsed" | "_coldAirportToday" | "_coldOne" | "_communismUsed" | "_confusingLEDClockUsed" | "_controlPanelUsed" | "_corruptedStardustUsed" | "_cosmicSixPackConjured" | "_crappyCameraUsed" | "_creepyVoodooDollUsed" | "_crimboTree" | "_cursedKegUsed" | "_cursedMicrowaveUsed" | "_dailyDungeonMalwareUsed" | "_darkChocolateHeart" | "_daycareFights" | "_daycareNap" | "_daycareSpa" | "_daycareToday" | "_defectiveTokenChecked" | "_defectiveTokenUsed" | "_dinseyGarbageDisposed" | "_discoKnife" | "_distentionPillUsed" | "_dnaHybrid" | "_drippingHallDoor1" | "_drippingHallDoor2" | "_drippingHallDoor3" | "_drippingHallDoor4" | "_drippyCaviarUsed" | "_drippyNuggetUsed" | "_drippyPilsnerUsed" | "_drippyPlumUsed" | "_drippyWineUsed" | "_eldritchHorrorEvoked" | "_eldritchTentacleFought" | "_envyfishEggUsed" | "_essentialTofuUsed" | "_etchedHourglassUsed" | "_eternalCarBatteryUsed" | "_everfullGlassUsed" | "_eyeAndATwistUsed" | "_fancyChessSetUsed" | "_falloutShelterSpaUsed" | "_fancyHotDogEaten" | "_farmerItemsCollected" | "_favoriteBirdVisited" | "_firedJokestersGun" | "_fireStartingKitUsed" | "_fireworksShop" | "_fireworksShopHatBought" | "_fireworksShopEquipmentBought" | "_fireworkUsed" | "_fishyPipeUsed" | "_floundryItemCreated" | "_floundryItemUsed" | "_freePillKeeperUsed" | "_frToday" | "_fudgeSporkUsed" | "_garbageItemChanged" | "_gingerBiggerAlligators" | "_gingerbreadCityToday" | "_gingerbreadClockAdvanced" | "_gingerbreadClockVisited" | "_gingerbreadColumnDestroyed" | "_gingerbreadMobHitUsed" | "_glennGoldenDiceUsed" | "_glitchItemImplemented" | "_gnollEyeUsed" | "_grimBuff" | "_guildManualUsed" | "_guzzlrQuestAbandoned" | "_hardKnocksDiplomaUsed" | "_hippyMeatCollected" | "_hobbyHorseUsed" | "_holidayFunUsed" | "_holoWristCrystal" | "_hotAirportToday" | "_hungerSauceUsed" | "_hyperinflatedSealLungUsed" | "_iceHotelRoomsRaided" | "_iceSculptureUsed" | "_incredibleSelfEsteemCast" | "_infernoDiscoVisited" | "_internetDailyDungeonMalwareBought" | "_internetGallonOfMilkBought" | "_internetPlusOneBought" | "_internetPrintScreenButtonBought" | "_internetViralVideoBought" | "_interviewIsabella" | "_interviewMasquerade" | "_interviewVlad" | "_ironicMoustache" | "_jackassPlumberGame" | "_jarlsCheeseSummoned" | "_jarlsCreamSummoned" | "_jarlsDoughSummoned" | "_jarlsEggsSummoned" | "_jarlsFruitSummoned" | "_jarlsMeatSummoned" | "_jarlsPotatoSummoned" | "_jarlsVeggiesSummoned" | "_jingleBellUsed" | "_jukebox" | "_kgbFlywheelCharged" | "_kgbLeftDrawerUsed" | "_kgbOpened" | "_kgbRightDrawerUsed" | "_kolConSixPackUsed" | "_kolhsCutButNotDried" | "_kolhsIsskayLikeAnAshtray" | "_kolhsPoeticallyLicenced" | "_kolhsSchoolSpirited" | "_latteBanishUsed" | "_latteCopyUsed" | "_latteDrinkUsed" | "_legendaryBeat" | "_licenseToChillUsed" | "_lookingGlass" | "_loveTunnelUsed" | "_luckyGoldRingVolcoino" | "_lunchBreak" | "_lupineHormonesUsed" | "_lyleFavored" | "_madTeaParty" | "_mafiaMiddleFingerRingUsed" | "_managerialManipulationUsed" | "_mayoDeviceRented" | "_mayoTankSoaked" | "_milkOfMagnesiumUsed" | "_mimeArmyShotglassUsed" | "_missileLauncherUsed" | "_momFoodReceived" | "_muffinOrderedToday" | "_mushroomGardenVisited" | "_neverendingPartyToday" | "_newYouQuestCompleted" | "_olympicSwimmingPool" | "_olympicSwimmingPoolItemFound" | "_overflowingGiftBasketUsed" | "_partyHard" | "_pastaAdditive" | "_perfectFreezeUsed" | "_perfectlyFairCoinUsed" | "_petePartyThrown" | "_peteRiotIncited" | "_photocopyUsed" | "_pickyTweezersUsed" | "_pirateBellowUsed" | "_pirateForkUsed" | "_pixelOrbUsed" | "_pneumaticityPotionUsed" | "_pottedTeaTreeUsed" | "_prToday" | "_psychoJarFilled" | "_psychoJarUsed" | "_psychokineticHugUsed" | "_rainStickUsed" | "_redwoodRainStickUsed" | "_requestSandwichSucceeded" | "_rhinestonesAcquired" | "_seaJellyHarvested" | "_setOfJacksUsed" | "_sewingKitUsed" | "_sexChanged" | "_shrubDecorated" | "_silverDreadFlaskUsed" | "_skateBuff1" | "_skateBuff2" | "_skateBuff3" | "_skateBuff4" | "_skateBuff5" | "_sleazeAirportToday" | "_sobrieTeaUsed" | "_softwareGlitchTurnReceived" | "_spacegateMurderbot" | "_spacegateRuins" | "_spacegateSpant" | "_spacegateToday" | "_spacegateVaccine" | "_spaghettiBreakfast" | "_spaghettiBreakfastEaten" | "_spinmasterLatheVisited" | "_spinningWheel" | "_spookyAirportToday" | "_stabonicScrollUsed" | "_steelyEyedSquintUsed" | "_stenchAirportToday" | "_stinkyCheeseBanisherUsed" | "_streamsCrossed" | "_stuffedPocketwatchUsed" | "_styxSprayUsed" | "_summonAnnoyanceUsed" | "_summonCarrotUsed" | "_summonResortPassUsed" | "_sweetToothUsed" | "_syntheticDogHairPillUsed" | "_tacoFlierUsed" | "_templeHiddenPower" | "_tempuraAirUsed" | "_thesisDelivered" | "_timeSpinnerReplicatorUsed" | "_toastSummoned" | "_tonicDjinn" | "_treasuryEliteMeatCollected" | "_treasuryHaremMeatCollected" | "_trivialAvocationsGame" | "_tryptophanDartUsed" | "_turtlePowerCast" | "_twelveNightEnergyUsed" | "_ultraMegaSourBallUsed" | "_universalSeasoningUsed" | "_victorSpoilsUsed" | "_villainLairCanLidUsed" | "_villainLairColorChoiceUsed" | "_villainLairDoorChoiceUsed" | "_villainLairFirecrackerUsed" | "_villainLairSymbologyChoiceUsed" | "_villainLairWebUsed" | "_vmaskBanisherUsed" | "_voraciTeaUsed" | "_volcanoItemRedeemed" | "_volcanoSuperduperheatedMetal" | "_voteToday" | "_VYKEACafeteriaRaided" | "_VYKEALoungeRaided" | "_walfordQuestStartedToday" | "_warbearBankUsed" | "_warbearBreakfastMachineUsed" | "_warbearGyrocopterUsed" | "_warbearSodaMachineUsed" | "_wildfireBarrelHarvested" | "_witchessBuff" | "_workshedItemUsed" | "_zombieClover" | "_preventScurvy" | "lockedItem4637" | "lockedItem4638" | "lockedItem4639" | "lockedItem4646" | "lockedItem4647" | "unknownRecipe3542" | "unknownRecipe3543" | "unknownRecipe3544" | "unknownRecipe3545" | "unknownRecipe3546" | "unknownRecipe3547" | "unknownRecipe3548" | "unknownRecipe3749" | "unknownRecipe3751" | "unknownRecipe4172" | "unknownRecipe4173" | "unknownRecipe4174" | "unknownRecipe5060" | "unknownRecipe5061" | "unknownRecipe5062" | "unknownRecipe5063" | "unknownRecipe5064" | "unknownRecipe5066" | "unknownRecipe5067" | "unknownRecipe5069" | "unknownRecipe5070" | "unknownRecipe5072" | "unknownRecipe5073" | "unknownRecipe5670" | "unknownRecipe5671" | "unknownRecipe6501" | "unknownRecipe6564" | "unknownRecipe6565" | "unknownRecipe6566" | "unknownRecipe6567" | "unknownRecipe6568" | "unknownRecipe6569" | "unknownRecipe6570" | "unknownRecipe6571" | "unknownRecipe6572" | "unknownRecipe6573" | "unknownRecipe6574" | "unknownRecipe6575" | "unknownRecipe6576" | "unknownRecipe6577" | "unknownRecipe6578" | "unknownRecipe7752" | "unknownRecipe7753" | "unknownRecipe7754" | "unknownRecipe7755" | "unknownRecipe7756" | "unknownRecipe7757" | "unknownRecipe7758";
+export type NumericProperty = "charsheetDropdown" | "chatStyle" | "coinMasterIndex" | "dailyDeedsVersion" | "defaultDropdown1" | "defaultDropdown2" | "defaultDropdownSplit" | "defaultLimit" | "fixedThreadPoolSize" | "itemManagerIndex" | "lastBuffRequestType" | "lastGlobalCounterDay" | "lastImageCacheClear" | "lastRssUpdate" | "previousUpdateRevision" | "relaySkillButtonCount" | "scriptButtonPosition" | "statusDropdown" | "svnThreadPoolSize" | "toolbarPosition" | "_g9Effect" | "addingScrolls" | "affirmationCookiesEaten" | "aminoAcidsUsed" | "antagonisticSnowmanKitCost" | "autoAbortThreshold" | "autoAntidote" | "autoBuyPriceLimit" | "availableCandyCredits" | "availableDimes" | "availableFunPoints" | "availableQuarters" | "availableStoreCredits" | "availableSwagger" | "averageSwagger" | "awolMedicine" | "awolPointsBeanslinger" | "awolPointsCowpuncher" | "awolPointsSnakeoiler" | "awolDeferredPointsBeanslinger" | "awolDeferredPointsCowpuncher" | "awolDeferredPointsSnakeoiler" | "awolVenom" | "bagOTricksCharges" | "ballpitBonus" | "bankedKarma" | "barrelGoal" | "bartenderTurnsUsed" | "basementMallPrices" | "basementSafetyMargin" | "batmanFundsAvailable" | "batmanBonusInitialFunds" | "batmanTimeLeft" | "bearSwagger" | "beeCounter" | "birdformCold" | "birdformHot" | "birdformRoc" | "birdformSleaze" | "birdformSpooky" | "birdformStench" | "blackBartsBootyCost" | "blackPuddingsDefeated" | "blackForestProgress" | "blankOutUsed" | "bloodweiserDrunk" | "bondPoints" | "bondVillainsDefeated" | "boneAbacusVictories" | "booPeakProgress" | "borisPoints" | "breakableHandling" | "breakableHandling1964" | "breakableHandling9691" | "breakableHandling9692" | "breakableHandling9699" | "brodenBacteria" | "brodenSprinkles" | "buffBotMessageDisposal" | "buffBotPhilanthropyType" | "buffJimmyIngredients" | "burnoutsDefeated" | "burrowgrubSummonsRemaining" | "camelSpit" | "camerasUsed" | "campAwayDecoration" | "carboLoading" | "catBurglarBankHeists" | "cellarLayout" | "charitableDonations" | "chasmBridgeProgress" | "chefTurnsUsed" | "chessboardsCleared" | "chilledToTheBone" | "cinderellaMinutesToMidnight" | "cinderellaScore" | "cocktailSummons" | "controlPanelOmega" | "cornucopiasOpened" | "cozyCounter6332" | "cozyCounter6333" | "cozyCounter6334" | "crimbo16BeardChakraCleanliness" | "crimbo16BootsChakraCleanliness" | "crimbo16BungChakraCleanliness" | "crimbo16CrimboHatChakraCleanliness" | "crimbo16GutsChakraCleanliness" | "crimbo16HatChakraCleanliness" | "crimbo16JellyChakraCleanliness" | "crimbo16LiverChakraCleanliness" | "crimbo16NippleChakraCleanliness" | "crimbo16NoseChakraCleanliness" | "crimbo16ReindeerChakraCleanliness" | "crimbo16SackChakraCleanliness" | "crimboTreeDays" | "cubelingProgress" | "currentExtremity" | "currentHedgeMazeRoom" | "currentMojoFilters" | "currentNunneryMeat" | "cyrptAlcoveEvilness" | "cyrptCrannyEvilness" | "cyrptNicheEvilness" | "cyrptNookEvilness" | "cyrptTotalEvilness" | "darkGyfftePoints" | "daycareEquipment" | "daycareInstructors" | "daycareLastScavenge" | "daycareToddlers" | "dbNemesisSkill1" | "dbNemesisSkill2" | "dbNemesisSkill3" | "desertExploration" | "desktopHeight" | "desktopWidth" | "dinseyFilthLevel" | "dinseyFunProgress" | "dinseyNastyBearsDefeated" | "dinseySocialJusticeIProgress" | "dinseySocialJusticeIIProgress" | "dinseyTouristsFed" | "dinseyToxicMultiplier" | "doctorBagQuestLights" | "doctorBagUpgrades" | "dreadScroll1" | "dreadScroll2" | "dreadScroll3" | "dreadScroll4" | "dreadScroll5" | "dreadScroll6" | "dreadScroll7" | "dreadScroll8" | "dripAdventuresSinceAscension" | "drippingHallAdventuresSinceAscension" | "drippingTreesAdventuresSinceAscension" | "drippyBatsUnlocked" | "drippyJuice" | "drippyOrbsClaimed" | "drunkenSwagger" | "edDefeatAbort" | "edPoints" | "eldritchTentaclesFought" | "electricKoolAidEaten" | "encountersUntilDMTChoice" | "ensorceleeLevel" | "essenceOfAnnoyanceCost" | "essenceOfBearCost" | "extraRolloverAdventures" | "falloutShelterLevel" | "fingernailsClipped" | "fistSkillsKnown" | "flyeredML" | "fossilB" | "fossilD" | "fossilN" | "fossilP" | "fossilS" | "fossilW" | "fratboysDefeated" | "frenchGuardTurtlesFreed" | "garbageChampagneCharge" | "garbageFireProgress" | "garbageShirtCharge" | "garbageTreeCharge" | "garlandUpgrades" | "gingerDigCount" | "gingerLawChoice" | "gingerMuscleChoice" | "gingerTrainScheduleStudies" | "gladiatorBallMovesKnown" | "gladiatorBladeMovesKnown" | "gladiatorNetMovesKnown" | "glitchItemCost" | "glitchItemImplementationCount" | "glitchItemImplementationLevel" | "glitchSwagger" | "gloverPoints" | "gnasirProgress" | "goldenMrAccessories" | "gongPath" | "goreCollected" | "gourdItemCount" | "grimoire1Summons" | "grimoire2Summons" | "grimoire3Summons" | "grimstoneCharge" | "guardTurtlesFreed" | "guideToSafariCost" | "guyMadeOfBeesCount" | "guzzlrBronzeDeliveries" | "guzzlrDeliveryProgress" | "guzzlrGoldDeliveries" | "guzzlrPlatinumDeliveries" | "haciendaLayout" | "heavyRainsStartingThunder" | "heavyRainsStartingRain" | "heavyRainsStartingLightning" | "heroDonationBoris" | "heroDonationJarlsberg" | "heroDonationSneakyPete" | "hiddenApartmentProgress" | "hiddenBowlingAlleyProgress" | "hiddenHospitalProgress" | "hiddenOfficeProgress" | "hiddenTavernUnlock" | "highTopPumped" | "hippiesDefeated" | "holidayHalsBookCost" | "holidaySwagger" | "hpAutoRecovery" | "hpAutoRecoveryTarget" | "iceSwagger" | "item9084" | "jarlsbergPoints" | "jungCharge" | "junglePuns" | "knownAscensions" | "kolhsTotalSchoolSpirited" | "lastAnticheeseDay" | "lastArcadeAscension" | "lastBadMoonReset" | "lastBangPotionReset" | "lastBarrelSmashed" | "lastBattlefieldReset" | "lastBreakfast" | "lastCastleGroundUnlock" | "lastCastleTopUnlock" | "lastCellarReset" | "lastChanceThreshold" | "lastChasmReset" | "lastColosseumRoundWon" | "lastCouncilVisit" | "lastCounterDay" | "lastDesertUnlock" | "lastDispensaryOpen" | "lastDMTDuplication" | "lastDwarfFactoryReset" | "lastEVHelmetValue" | "lastEVHelmetReset" | "lastEasterEggBalloon" | "lastEmptiedStorage" | "lastFilthClearance" | "lastGoofballBuy" | "lastGuildStoreOpen" | "lastGuyMadeOfBeesReset" | "lastFratboyCall" | "lastFriarCeremonyAscension" | "lastHippyCall" | "lastIslandUnlock" | "lastKeyotronUse" | "lastKingLiberation" | "lastLightsOutTurn" | "lastMushroomPlot" | "lastMiningReset" | "lastNemesisReset" | "lastPaperStripReset" | "lastPirateEphemeraReset" | "lastPirateInsultReset" | "lastPlusSignUnlock" | "lastQuartetAscension" | "lastQuartetRequest" | "lastSecondFloorUnlock" | "lastSemirareReset" | "lastSkateParkReset" | "lastStillBeatingSpleen" | "lastTavernAscension" | "lastTavernSquare" | "lastTelescopeReset" | "lastTempleAdventures" | "lastTempleButtonsUnlock" | "lastTempleUnlock" | "lastTr4pz0rQuest" | "lastVioletFogMap" | "lastVoteMonsterTurn" | "lastWartDinseyDefeated" | "lastWuTangDefeated" | "lastYearbookCameraAscension" | "lastZapperWand" | "lawOfAveragesCost" | "libramSummons" | "lightsOutAutomation" | "louvreDesiredGoal" | "louvreGoal" | "lttQuestDifficulty" | "lttQuestStageCount" | "manaBurnSummonThreshold" | "manaBurningThreshold" | "manaBurningTrigger" | "manorDrawerCount" | "manualOfNumberologyCost" | "mapToKokomoCost" | "masksUnlocked" | "maximizerMRUSize" | "maximizerCombinationLimit" | "maximizerEquipmentLevel" | "maximizerEquipmentScope" | "maximizerMaxPrice" | "maximizerPriceLevel" | "maxManaBurn" | "mayflyExperience" | "mayoLevel" | "meansuckerPrice" | "merkinVocabularyMastery" | "miniAdvClass" | "miniMartinisDrunk" | "moleTunnelLevel" | "mothershipProgress" | "mpAutoRecovery" | "mpAutoRecoveryTarget" | "munchiesPillsUsed" | "mushroomGardenCropLevel" | "nextParanormalActivity" | "nextQuantumFamiliarTurn" | "noobPoints" | "noobDeferredPoints" | "noodleSummons" | "nsContestants1" | "nsContestants2" | "nsContestants3" | "numericSwagger" | "nunsVisits" | "oilPeakProgress" | "optimalSwagger" | "optimisticCandleProgress" | "palindomeDudesDefeated" | "parasolUsed" | "pendingMapReflections" | "pirateSwagger" | "plantingDay" | "plumberBadgeCost" | "plumberCostumeCost" | "plumberPoints" | "poolSharkCount" | "poolSkill" | "prismaticSummons" | "procrastinatorLanguageFluency" | "promptAboutCrafting" | "puzzleChampBonus" | "pyramidPosition" | "rockinRobinProgress" | "ROMOfOptimalityCost" | "quantumPoints" | "reagentSummons" | "reanimatorArms" | "reanimatorLegs" | "reanimatorSkulls" | "reanimatorWeirdParts" | "reanimatorWings" | "recentLocations" | "redSnapperProgress" | "relocatePygmyJanitor" | "relocatePygmyLawyer" | "rumpelstiltskinTurnsUsed" | "rumpelstiltskinKidsRescued" | "safariSwagger" | "sausageGrinderUnits" | "schoolOfHardKnocksDiplomaCost" | "schoolSwagger" | "scrapbookCharges" | "scriptMRULength" | "seaodesFound" | "SeasoningSwagger" | "semirareCounter" | "sexChanges" | "shenInitiationDay" | "shockingLickCharges" | "singleFamiliarRun" | "skillBurn3" | "skillBurn90" | "skillBurn153" | "skillBurn154" | "skillBurn155" | "skillBurn1019" | "skillBurn5017" | "skillBurn6014" | "skillBurn6015" | "skillBurn6016" | "skillBurn6020" | "skillBurn6021" | "skillBurn6022" | "skillBurn6023" | "skillBurn6024" | "skillBurn6026" | "skillBurn6028" | "skillBurn7323" | "skillBurn14008" | "skillBurn14028" | "skillBurn14038" | "skillBurn15011" | "skillBurn15028" | "skillBurn17005" | "skillBurn22034" | "skillBurn22035" | "skillBurn23301" | "skillBurn23302" | "skillBurn23303" | "skillBurn23304" | "skillBurn23305" | "skillBurn23306" | "skillLevel46" | "skillLevel47" | "skillLevel48" | "skillLevel117" | "skillLevel118" | "skillLevel121" | "skillLevel128" | "skillLevel134" | "skillLevel144" | "skillLevel180" | "skillLevel188" | "skillLevel7254" | "slimelingFullness" | "slimelingStacksDropped" | "slimelingStacksDue" | "smoresEaten" | "smutOrcNoncombatProgress" | "sneakyPetePoints" | "snojoMoxieWins" | "snojoMuscleWins" | "snojoMysticalityWins" | "sourceAgentsDefeated" | "sourceEnlightenment" | "sourceInterval" | "sourcePoints" | "sourceTerminalGram" | "sourceTerminalPram" | "sourceTerminalSpam" | "spaceBabyLanguageFluency" | "spacePirateLanguageFluency" | "spelunkyNextNoncombat" | "spelunkySacrifices" | "spelunkyWinCount" | "spookyPuttyCopiesMade" | "statbotUses" | "sugarCounter4178" | "sugarCounter4179" | "sugarCounter4180" | "sugarCounter4181" | "sugarCounter4182" | "sugarCounter4183" | "sugarCounter4191" | "summonAnnoyanceCost" | "tacoDanCocktailSauce" | "tacoDanFishMeat" | "tavernLayout" | "telescopeUpgrades" | "tempuraSummons" | "timeSpinnerMedals" | "timesRested" | "tomeSummons" | "totalCharitableDonations" | "turtleBlessingTurns" | "twinPeakProgress" | "unicornHornInflation" | "universalSeasoningCost" | "usable1HWeapons" | "usable1xAccs" | "usable2HWeapons" | "usable3HWeapons" | "usableAccessories" | "usableHats" | "usableOffhands" | "usableOther" | "usablePants" | "usableShirts" | "valueOfAdventure" | "valueOfInventory" | "valueOfStill" | "valueOfTome" | "violetFogGoal" | "walfordBucketProgress" | "warehouseProgress" | "welcomeBackAdv" | "writingDesksDefeated" | "xoSkeleltonXProgress" | "xoSkeleltonOProgress" | "yearbookCameraAscensions" | "yearbookCameraUpgrades" | "youRobotBody" | "youRobotBottom" | "youRobotLeft" | "youRobotRight" | "youRobotTop" | "zeppelinProtestors" | "zombiePoints" | "_absintheDrops" | "_abstractionDropsCrown" | "_aguaDrops" | "_xenomorphCharge" | "_ancestralRecallCasts" | "_antihangoverBonus" | "_astralDrops" | "_backUpUses" | "_badlyRomanticArrows" | "_badgerCharge" | "_balefulHowlUses" | "_banderRunaways" | "_bastilleGames" | "_beanCannonUses" | "_bearHugs" | "_beerLensDrops" | "_benettonsCasts" | "_birdsSoughtToday" | "_boomBoxFights" | "_boomBoxSongsLeft" | "_bootStomps" | "_boxingGloveArrows" | "_brickoEyeSummons" | "_brickoFights" | "_campAwayCloudBuffs" | "_campAwaySmileBuffs" | "_candySummons" | "_carnieCandyDrops" | "_carrotNoseDrops" | "_catBurglarCharge" | "_catBurglarHeistsComplete" | "_cheerleaderSteam" | "_chestXRayUsed" | "_chipBags" | "_chocolateCigarsUsed" | "_chocolateSculpturesUsed" | "_chocolatesUsed" | "_chronolithActivations" | "_clanFortuneConsultUses" | "_clipartSummons" | "_companionshipCasts" | "_dailySpecialPrice" | "_daycareGymScavenges" | "_daycareRecruits" | "_deckCardsDrawn" | "_deluxeKlawSummons" | "_demandSandwich" | "_detectiveCasesCompleted" | "_disavowed" | "_dnaPotionsMade" | "_donhosCasts" | "_dreamJarDrops" | "_drunkPygmyBanishes" | "_edDefeats" | "_edLashCount" | "_elronsCasts" | "_enamorangs" | "_energyCollected" | "_expertCornerCutterUsed" | "_favorRareSummons" | "_feastUsed" | "_feelinTheRhythm" | "_feelPrideUsed" | "_feelExcitementUsed" | "_feelHatredUsed" | "_feelLonelyUsed" | "_feelNervousUsed" | "_feelEnvyUsed" | "_feelDisappointedUsed" | "_feelSuperiorUsed" | "_feelLostUsed" | "_feelNostalgicUsed" | "_feelPeacefulUsed" | "_fingertrapArrows" | "_fragrantHerbsUsed" | "_freeBeachWalksUsed" | "_frButtonsPressed" | "_fudgeWaspFights" | "_gapBuffs" | "_garbageFireDropsCrown" | "_genieFightsUsed" | "_genieWishesUsed" | "_gibbererAdv" | "_gibbererCharge" | "_gingerbreadCityTurns" | "_glarkCableUses" | "_glitchMonsterFights" | "_gnomeAdv" | "_godLobsterFights" | "_goldenMoneyCharge" | "_gongDrops" | "_gothKidCharge" | "_gothKidFights" | "_grimBrotherCharge" | "_grimFairyTaleDrops" | "_grimFairyTaleDropsCrown" | "_grimoireConfiscatorSummons" | "_grimoireGeekySummons" | "_grimstoneMaskDrops" | "_grimstoneMaskDropsCrown" | "_grooseCharge" | "_grooseDrops" | "_guzzlrDeliveries" | "_guzzlrGoldDeliveries" | "_guzzlrPlatinumDeliveries" | "_hareAdv" | "_hareCharge" | "_highTopPumps" | "_hipsterAdv" | "_hoardedCandyDropsCrown" | "_hoboUnderlingSummons" | "_holoWristDrops" | "_holoWristProgress" | "_hotAshesDrops" | "_hotJellyUses" | "_hotTubSoaks" | "_humanMuskUses" | "_iceballUses" | "_inigosCasts" | "_jerksHealthMagazinesUsed" | "_jiggleCheese" | "_jiggleCream" | "_jiggleLife" | "_jiggleSteak" | "_jitbCharge" | "_jungDrops" | "_kgbClicksUsed" | "_kgbDispenserUses" | "_kgbTranquilizerDartUses" | "_klawSummons" | "_kloopCharge" | "_kloopDrops" | "_kolhsAdventures" | "_kolhsSavedByTheBell" | "_lastDailyDungeonRoom" | "_lastSausageMonsterTurn" | "_lastZomboEye" | "_latteRefillsUsed" | "_leafblowerML" | "_legionJackhammerCrafting" | "_llamaCharge" | "_longConUsed" | "_loveChocolatesUsed" | "_lynyrdSnareUses" | "_machineTunnelsAdv" | "_macrometeoriteUses" | "_mafiaThumbRingAdvs" | "_monstersMapped" | "_mayflowerDrops" | "_mayflySummons" | "_mediumSiphons" | "_meteoriteAdesUsed" | "_meteorShowerUses" | "_micrometeoriteUses" | "_miniMartiniDrops" | "_mushroomGardenFights" | "_nanorhinoCharge" | "_navelRunaways" | "_neverendingPartyFreeTurns" | "_newYouQuestSharpensDone" | "_newYouQuestSharpensToDo" | "_nextQuantumAlignment" | "_nightmareFuelCharges" | "_noobSkillCount" | "_nuclearStockpileUsed" | "_oilExtracted" | "_optimisticCandleDropsCrown" | "_oreDropsCrown" | "_otoscopeUsed" | "_pantsgivingBanish" | "_pantsgivingCount" | "_pantsgivingCrumbs" | "_pantsgivingFullness" | "_pasteDrops" | "_peteJukeboxFixed" | "_peteJumpedShark" | "_petePeeledOut" | "_pieDrops" | "_piePartsCount" | "_pixieCharge" | "_pocketProfessorLectures" | "_poisonArrows" | "_pokeGrowFertilizerDrops" | "_poolGames" | "_powderedGoldDrops" | "_powderedMadnessUses" | "_powerfulGloveBatteryPowerUsed" | "_powerPillDrops" | "_powerPillUses" | "_precisionCasts" | "_radlibSummons" | "_raindohCopiesMade" | "_rapidPrototypingUsed" | "_raveStealCount" | "_reflexHammerUsed" | "_resolutionAdv" | "_resolutionRareSummons" | "_riftletAdv" | "_rogueProgramCharge" | "_romanticFightsLeft" | "_saberForceMonsterCount" | "_saberForceUses" | "_saberMod" | "_saltGrainsConsumed" | "_sandwormCharge" | "_saplingsPlanted" | "_sausageFights" | "_sausagesEaten" | "_sausagesMade" | "_sealFigurineUses" | "_sealScreeches" | "_sealsSummoned" | "_shatteringPunchUsed" | "_shortOrderCookCharge" | "_shrubCharge" | "_sloppyDinerBeachBucks" | "_smilesOfMrA" | "_smithsnessSummons" | "_snojoFreeFights" | "_snojoParts" | "_snokebombUsed" | "_snowconeSummons" | "_snowglobeDrops" | "_snowSuitCount" | "_sourceTerminalDigitizeMonsterCount" | "_sourceTerminalDigitizeUses" | "_sourceTerminalDuplicateUses" | "_sourceTerminalEnhanceUses" | "_sourceTerminalExtrudes" | "_sourceTerminalPortscanUses" | "_spaceFurDropsCrown" | "_spacegatePlanetIndex" | "_spacegateTurnsLeft" | "_spaceJellyfishDrops" | "_speakeasyDrinksDrunk" | "_spelunkerCharges" | "_spelunkingTalesDrops" | "_spookyJellyUses" | "_stackLumpsUses" | "_steamCardDrops" | "_stickerSummons" | "_stinkyCheeseCount" | "_stressBallSqueezes" | "_sugarSummons" | "_taffyRareSummons" | "_taffyYellowSummons" | "_thanksgettingFoodsEaten" | "_thingfinderCasts" | "_thinknerdPackageDrops" | "_thorsPliersCrafting" | "_timeHelmetAdv" | "_timeSpinnerMinutesUsed" | "_tokenDrops" | "_transponderDrops" | "_turkeyBlastersUsed" | "_turkeyBooze" | "_turkeyMuscle" | "_turkeyMyst" | "_turkeyMoxie" | "_unaccompaniedMinerUsed" | "_unconsciousCollectiveCharge" | "_universeCalculated" | "_universeImploded" | "_usedReplicaBatoomerang" | "_vampyreCloakeFormUses" | "_villainLairProgress" | "_vitachocCapsulesUsed" | "_vmaskAdv" | "_volcanoItem1" | "_volcanoItem2" | "_volcanoItem3" | "_volcanoItemCount1" | "_volcanoItemCount2" | "_volcanoItemCount3" | "_voteFreeFights" | "_VYKEACompanionLevel" | "_warbearAutoAnvilCrafting" | "_whiteRiceDrops" | "_witchessFights" | "_xoHugsUsed" | "_yellowPixelDropsCrown" | "_zapCount";
+export type MonsterProperty = "cameraMonster" | "chateauMonster" | "crappyCameraMonster" | "crudeMonster" | "crystalBallMonster" | "enamorangMonster" | "envyfishMonster" | "lastCopyableMonster" | "iceSculptureMonster" | "longConMonster" | "makeFriendsMonster" | "merkinLockkeyMonster" | "nosyNoseMonster" | "olfactedMonster" | "photocopyMonster" | "rainDohMonster" | "romanticTarget" | "screencappedMonster" | "spookyPuttyMonster" | "stenchCursedMonster" | "superficiallyInterestedMonster" | "waxMonster" | "yearbookCameraTarget" | "_gallapagosMonster" | "_jiggleCreamedMonster" | "_latteMonster" | "_nanorhinoBanishedMonster" | "_newYouQuestMonster" | "_relativityMonster" | "_saberForceMonster" | "_sourceTerminalDigitizeMonster" | "_voteMonster";
+export type LocationProperty = "crystalBallLocation" | "currentJunkyardLocation" | "doctorBagQuestLocation" | "ghostLocation" | "guzzlrQuestLocation" | "nextSpookyravenElizabethRoom" | "nextSpookyravenStephenRoom" | "semirareLocation" | "sourceOracleTarget";
+export type StringProperty = "autoLogin" | "browserBookmarks" | "chatFontSize" | "combatHotkey0" | "combatHotkey1" | "combatHotkey2" | "combatHotkey3" | "combatHotkey4" | "combatHotkey5" | "combatHotkey6" | "combatHotkey7" | "combatHotkey8" | "combatHotkey9" | "commandLineNamespace" | "cookies.inventory" | "dailyDeedsOptions" | "defaultBorderColor" | "displayName" | "externalEditor" | "getBreakfast" | "headerStates" | "highlightList" | "http.proxyHost" | "http.proxyPassword" | "http.proxyPort" | "http.proxyUser" | "https.proxyHost" | "https.proxyPassword" | "https.proxyPort" | "https.proxyUser" | "initialDesktop" | "initialFrames" | "innerChatColor" | "innerTabColor" | "lastRelayUpdate" | "lastRssVersion" | "lastUserAgent" | "lastUsername" | "logPreferenceChangeFilter" | "loginScript" | "loginServerName" | "loginWindowLogo" | "logoutScript" | "outerChatColor" | "outerTabColor" | "preferredWebBrowser" | "previousNotifyList" | "previousUpdateVersion" | "saveState" | "saveStateActive" | "scriptList" | "swingLookAndFeel" | "useDecoratedTabs" | "userAgent" | "afterAdventureScript" | "autoOlfact" | "autoPutty" | "backupCameraMode" | "banishedMonsters" | "banishingShoutMonsters" | "barrelLayout" | "batmanStats" | "batmanZone" | "batmanUpgrades" | "battleAction" | "beachHeadsUnlocked" | "beforePVPScript" | "betweenBattleScript" | "boomBoxSong" | "breakfastAlways" | "breakfastHardcore" | "breakfastSoftcore" | "buffBotCasting" | "buyScript" | "cargoPocketsEmptied" | "cargoPocketScraps" | "chatbotScript" | "chatPlayerScript" | "choiceAdventureScript" | "chosenTrip" | "clanFortuneReply1" | "clanFortuneReply2" | "clanFortuneReply3" | "clanFortuneWord1" | "clanFortuneWord2" | "clanFortuneWord3" | "commerceGhostItem" | "counterScript" | "copperheadClubHazard" | "crimbotChassis" | "crimbotArm" | "crimbotPropulsion" | "csServicesPerformed" | "currentEasyBountyItem" | "currentHardBountyItem" | "currentHippyStore" | "currentJunkyardTool" | "currentMood" | "currentPVPSeason" | "currentPvpVictories" | "currentSpecialBountyItem" | "customCombatScript" | "cyrusAdjectives" | "defaultFlowerLossMessage" | "defaultFlowerWinMessage" | "demonName1" | "demonName2" | "demonName3" | "demonName4" | "demonName5" | "demonName6" | "demonName7" | "demonName8" | "demonName9" | "demonName10" | "demonName11" | "demonName12" | "demonName13" | "dinseyGatorStenchDamage" | "dinseyRollercoasterStats" | "dnaSyringe" | "doctorBagQuestItem" | "dolphinItem" | "edPiece" | "ensorcelee" | "EVEDirections" | "extraCosmeticModifiers" | "familiarScript" | "gameProBossSpecialPower" | "grimoireSkillsHardcore" | "grimoireSkillsSoftcore" | "grimstoneMaskPath" | "guzzlrQuestClient" | "guzzlrQuestBooze" | "guzzlrQuestTier" | "harvestGardenHardcore" | "harvestGardenSoftcore" | "hpAutoRecoveryItems" | "invalidBuffMessage" | "jickSwordModifier" | "kingLiberatedScript" | "lassoTraining" | "lastAdventure" | "lastBangPotion819" | "lastBangPotion820" | "lastBangPotion821" | "lastBangPotion822" | "lastBangPotion823" | "lastBangPotion824" | "lastBangPotion825" | "lastBangPotion826" | "lastBangPotion827" | "lastChanceBurn" | "lastChessboard" | "lastDwarfDiceRolls" | "lastDwarfDigitRunes" | "lastDwarfEquipmentRunes" | "lastDwarfFactoryItem118" | "lastDwarfFactoryItem119" | "lastDwarfFactoryItem120" | "lastDwarfFactoryItem360" | "lastDwarfFactoryItem361" | "lastDwarfFactoryItem362" | "lastDwarfFactoryItem363" | "lastDwarfFactoryItem364" | "lastDwarfFactoryItem365" | "lastDwarfFactoryItem910" | "lastDwarfFactoryItem3199" | "lastDwarfOfficeItem3208" | "lastDwarfOfficeItem3209" | "lastDwarfOfficeItem3210" | "lastDwarfOfficeItem3211" | "lastDwarfOfficeItem3212" | "lastDwarfOfficeItem3213" | "lastDwarfOfficeItem3214" | "lastDwarfOreRunes" | "lastDwarfHopper1" | "lastDwarfHopper2" | "lastDwarfHopper3" | "lastDwarfHopper4" | "lastEncounter" | "lastMacroError" | "lastMessageId" | "lastPaperStrip3144" | "lastPaperStrip4138" | "lastPaperStrip4139" | "lastPaperStrip4140" | "lastPaperStrip4141" | "lastPaperStrip4142" | "lastPaperStrip4143" | "lastPaperStrip4144" | "lastPirateEphemera" | "lastPorkoBoard" | "lastPorkoPayouts" | "lastPorkoExpected" | "lastSlimeVial3885" | "lastSlimeVial3886" | "lastSlimeVial3887" | "lastSlimeVial3888" | "lastSlimeVial3889" | "lastSlimeVial3890" | "lastSlimeVial3891" | "lastSlimeVial3892" | "lastSlimeVial3893" | "lastSlimeVial3894" | "lastSlimeVial3895" | "lastSlimeVial3896" | "latteModifier" | "latteUnlocks" | "libramSkillsHardcore" | "libramSkillsSoftcore" | "louvreOverride" | "lovePotion" | "lttQuestName" | "maximizerList" | "maximizerMRUList" | "mayoInMouth" | "mayoMinderSetting" | "merkinQuestPath" | "mineLayout1" | "mineLayout2" | "mineLayout3" | "mineLayout4" | "mineLayout5" | "mineLayout6" | "mpAutoRecoveryItems" | "muffinOnOrder" | "nextAdventure" | "nsChallenge2" | "nsChallenge3" | "nsChallenge4" | "nsChallenge5" | "nsTowerDoorKeysUsed" | "oceanAction" | "oceanDestination" | "pastaThrall1" | "pastaThrall2" | "pastaThrall3" | "pastaThrall4" | "pastaThrall5" | "pastaThrall6" | "pastaThrall7" | "pastaThrall8" | "peteMotorbikeTires" | "peteMotorbikeGasTank" | "peteMotorbikeHeadlight" | "peteMotorbikeCowling" | "peteMotorbikeMuffler" | "peteMotorbikeSeat" | "pieStuffing" | "plantingDate" | "plantingLength" | "plantingScript" | "plumberCostumeWorn" | "pokefamBoosts" | "postAscensionScript" | "preAscensionScript" | "retroCapeSuperhero" | "retroCapeWashingInstructions" | "questDoctorBag" | "questECoBucket" | "questESlAudit" | "questESlBacteria" | "questESlCheeseburger" | "questESlCocktail" | "questESlDebt" | "questESlFish" | "questESlMushStash" | "questESlSalt" | "questESlSprinkles" | "questESpEVE" | "questESpJunglePun" | "questESpGore" | "questESpClipper" | "questESpFakeMedium" | "questESpSerum" | "questESpSmokes" | "questESpOutOfOrder" | "questEStFishTrash" | "questEStGiveMeFuel" | "questEStNastyBears" | "questEStSocialJusticeI" | "questEStSocialJusticeII" | "questEStSuperLuber" | "questEStWorkWithFood" | "questEStZippityDooDah" | "questEUNewYou" | "questF01Primordial" | "questF02Hyboria" | "questF03Future" | "questF04Elves" | "questF05Clancy" | "questG01Meatcar" | "questG02Whitecastle" | "questG03Ego" | "questG04Nemesis" | "questG05Dark" | "questG06Delivery" | "questG07Myst" | "questG08Moxie" | "questG09Muscle" | "questGuzzlr" | "questI01Scapegoat" | "questI02Beat" | "questL02Larva" | "questL03Rat" | "questL04Bat" | "questL05Goblin" | "questL06Friar" | "questL07Cyrptic" | "questL08Trapper" | "questL09Topping" | "questL10Garbage" | "questL11MacGuffin" | "questL11Black" | "questL11Business" | "questL11Curses" | "questL11Desert" | "questL11Doctor" | "questL11Manor" | "questL11Palindome" | "questL11Pyramid" | "questL11Ron" | "questL11Shen" | "questL11Spare" | "questL11Worship" | "questL12War" | "questL12HippyFrat" | "questL13Final" | "questL13Warehouse" | "questLTTQuestByWire" | "questM01Untinker" | "questM02Artist" | "questM03Bugbear" | "questM05Toot" | "questM06Gourd" | "questM07Hammer" | "questM08Baker" | "questM09Rocks" | "questM10Azazel" | "questM11Postal" | "questM12Pirate" | "questM13Escape" | "questM14Bounty" | "questM15Lol" | "questM16Temple" | "questM17Babies" | "questM18Swamp" | "questM19Hippy" | "questM20Necklace" | "questM21Dance" | "questM22Shirt" | "questM23Meatsmith" | "questM24Doc" | "questM25Armorer" | "questM26Oracle" | "questPAGhost" | "questS01OldGuy" | "questS02Monkees" | "raveCombo1" | "raveCombo2" | "raveCombo3" | "raveCombo4" | "raveCombo5" | "raveCombo6" | "recoveryScript" | "redSnapperPhylum" | "relayCounters" | "royalty" | "scriptMRUList" | "seahorseName" | "shenQuestItem" | "shrubGarland" | "shrubGifts" | "shrubLights" | "shrubTopper" | "sideDefeated" | "sidequestArenaCompleted" | "sidequestFarmCompleted" | "sidequestJunkyardCompleted" | "sidequestLighthouseCompleted" | "sidequestNunsCompleted" | "sidequestOrchardCompleted" | "skateParkStatus" | "snowsuit" | "sourceTerminalChips" | "sourceTerminalEducate1" | "sourceTerminalEducate2" | "sourceTerminalEnquiry" | "sourceTerminalEducateKnown" | "sourceTerminalEnhanceKnown" | "sourceTerminalEnquiryKnown" | "sourceTerminalExtrudeKnown" | "spadingData" | "spadingScript" | "spelunkyStatus" | "spelunkyUpgrades" | "spookyravenRecipeUsed" | "stationaryButton1" | "stationaryButton2" | "stationaryButton3" | "stationaryButton4" | "stationaryButton5" | "streamCrossDefaultTarget" | "sweetSynthesisBlacklist" | "telescope1" | "telescope2" | "telescope3" | "telescope4" | "telescope5" | "testudinalTeachings" | "textColors" | "thanksMessage" | "tomeSkillsHardcore" | "tomeSkillsSoftcore" | "trackVoteMonster" | "trapperOre" | "umdLastObtained" | "violetFogLayout" | "volcanoMaze1" | "volcanoMaze2" | "volcanoMaze3" | "volcanoMaze4" | "volcanoMaze5" | "walfordBucketItem" | "warProgress" | "workteaClue" | "yourFavoriteBird" | "yourFavoriteBirdMods" | "youRobotCPUUpgrades" | "_beachHeadsUsed" | "_beachLayout" | "_beachMinutes" | "_birdOfTheDay" | "_birdOfTheDayMods" | "_bittycar" | "_campAwaySmileBuffSign" | "_cloudTalkMessage" | "_cloudTalkSmoker" | "_dailySpecial" | "_deckCardsSeen" | "_feastedFamiliars" | "_floristPlantsUsed" | "_frAreasUnlocked" | "_frHoursLeft" | "_frMonstersKilled" | "_horsery" | "_horseryCrazyMox" | "_horseryCrazyMus" | "_horseryCrazyMys" | "_horseryCrazyName" | "_horseryCurrentName" | "_horseryDarkName" | "_horseryNormalName" | "_horseryPaleName" | "_jickJarAvailable" | "_jiggleCheesedMonsters" | "_lastCombatStarted" | "_LastPirateRealmIsland" | "_mummeryMods" | "_mummeryUses" | "_newYouQuestSkill" | "_noHatModifier" | "_pantogramModifier" | "_questESp" | "_questPartyFair" | "_questPartyFairProgress" | "_questPartyFairQuest" | "_roboDrinks" | "_spacegateAnimalLife" | "_spacegateCoordinates" | "_spacegateHazards" | "_spacegateIntelligentLife" | "_spacegatePlanetName" | "_spacegatePlantLife" | "_stolenAccordions" | "_tempRelayCounters" | "_timeSpinnerFoodAvailable" | "_unknownEasyBountyItem" | "_unknownHardBountyItem" | "_unknownSpecialBountyItem" | "_untakenEasyBountyItem" | "_untakenHardBountyItem" | "_untakenSpecialBountyItem" | "_userMods" | "_villainLairColor" | "_villainLairKey" | "_voteLocal1" | "_voteLocal2" | "_voteLocal3" | "_voteLocal4" | "_voteMonster1" | "_voteMonster2" | "_voteModifier" | "_VYKEACompanionType" | "_VYKEACompanionRune" | "_VYKEACompanionName";
+export type NumericOrStringProperty = "statusEngineering" | "statusGalley" | "statusMedbay" | "statusMorgue" | "statusNavigation" | "statusScienceLab" | "statusSonar" | "statusSpecialOps" | "statusWasteProcessing" | "choiceAdventure2" | "choiceAdventure3" | "choiceAdventure4" | "choiceAdventure5" | "choiceAdventure6" | "choiceAdventure7" | "choiceAdventure8" | "choiceAdventure9" | "choiceAdventure10" | "choiceAdventure11" | "choiceAdventure12" | "choiceAdventure14" | "choiceAdventure15" | "choiceAdventure16" | "choiceAdventure17" | "choiceAdventure18" | "choiceAdventure19" | "choiceAdventure20" | "choiceAdventure21" | "choiceAdventure22" | "choiceAdventure23" | "choiceAdventure24" | "choiceAdventure25" | "choiceAdventure26" | "choiceAdventure27" | "choiceAdventure28" | "choiceAdventure29" | "choiceAdventure40" | "choiceAdventure41" | "choiceAdventure42" | "choiceAdventure45" | "choiceAdventure46" | "choiceAdventure47" | "choiceAdventure71" | "choiceAdventure72" | "choiceAdventure73" | "choiceAdventure74" | "choiceAdventure75" | "choiceAdventure76" | "choiceAdventure77" | "choiceAdventure86" | "choiceAdventure87" | "choiceAdventure88" | "choiceAdventure89" | "choiceAdventure90" | "choiceAdventure91" | "choiceAdventure105" | "choiceAdventure106" | "choiceAdventure107" | "choiceAdventure108" | "choiceAdventure109" | "choiceAdventure110" | "choiceAdventure111" | "choiceAdventure112" | "choiceAdventure113" | "choiceAdventure114" | "choiceAdventure115" | "choiceAdventure116" | "choiceAdventure117" | "choiceAdventure118" | "choiceAdventure120" | "choiceAdventure123" | "choiceAdventure125" | "choiceAdventure126" | "choiceAdventure127" | "choiceAdventure129" | "choiceAdventure131" | "choiceAdventure132" | "choiceAdventure135" | "choiceAdventure136" | "choiceAdventure137" | "choiceAdventure138" | "choiceAdventure139" | "choiceAdventure140" | "choiceAdventure141" | "choiceAdventure142" | "choiceAdventure143" | "choiceAdventure144" | "choiceAdventure145" | "choiceAdventure146" | "choiceAdventure147" | "choiceAdventure148" | "choiceAdventure149" | "choiceAdventure151" | "choiceAdventure152" | "choiceAdventure153" | "choiceAdventure154" | "choiceAdventure155" | "choiceAdventure156" | "choiceAdventure157" | "choiceAdventure158" | "choiceAdventure159" | "choiceAdventure160" | "choiceAdventure161" | "choiceAdventure162" | "choiceAdventure163" | "choiceAdventure164" | "choiceAdventure165" | "choiceAdventure166" | "choiceAdventure167" | "choiceAdventure168" | "choiceAdventure169" | "choiceAdventure170" | "choiceAdventure171" | "choiceAdventure172" | "choiceAdventure177" | "choiceAdventure178" | "choiceAdventure180" | "choiceAdventure181" | "choiceAdventure182" | "choiceAdventure184" | "choiceAdventure185" | "choiceAdventure186" | "choiceAdventure187" | "choiceAdventure188" | "choiceAdventure189" | "choiceAdventure191" | "choiceAdventure197" | "choiceAdventure198" | "choiceAdventure199" | "choiceAdventure200" | "choiceAdventure201" | "choiceAdventure202" | "choiceAdventure203" | "choiceAdventure204" | "choiceAdventure205" | "choiceAdventure206" | "choiceAdventure207" | "choiceAdventure208" | "choiceAdventure211" | "choiceAdventure212" | "choiceAdventure213" | "choiceAdventure214" | "choiceAdventure215" | "choiceAdventure216" | "choiceAdventure217" | "choiceAdventure218" | "choiceAdventure219" | "choiceAdventure220" | "choiceAdventure221" | "choiceAdventure222" | "choiceAdventure223" | "choiceAdventure224" | "choiceAdventure225" | "choiceAdventure272" | "choiceAdventure273" | "choiceAdventure276" | "choiceAdventure277" | "choiceAdventure278" | "choiceAdventure279" | "choiceAdventure280" | "choiceAdventure281" | "choiceAdventure282" | "choiceAdventure283" | "choiceAdventure284" | "choiceAdventure286" | "choiceAdventure287" | "choiceAdventure288" | "choiceAdventure289" | "choiceAdventure290" | "choiceAdventure291" | "choiceAdventure292" | "choiceAdventure293" | "choiceAdventure294" | "choiceAdventure295" | "choiceAdventure296" | "choiceAdventure297" | "choiceAdventure298" | "choiceAdventure299" | "choiceAdventure302" | "choiceAdventure303" | "choiceAdventure304" | "choiceAdventure305" | "choiceAdventure306" | "choiceAdventure307" | "choiceAdventure308" | "choiceAdventure309" | "choiceAdventure310" | "choiceAdventure311" | "choiceAdventure317" | "choiceAdventure318" | "choiceAdventure319" | "choiceAdventure320" | "choiceAdventure321" | "choiceAdventure322" | "choiceAdventure326" | "choiceAdventure327" | "choiceAdventure328" | "choiceAdventure329" | "choiceAdventure330" | "choiceAdventure331" | "choiceAdventure332" | "choiceAdventure333" | "choiceAdventure334" | "choiceAdventure335" | "choiceAdventure336" | "choiceAdventure337" | "choiceAdventure338" | "choiceAdventure339" | "choiceAdventure340" | "choiceAdventure341" | "choiceAdventure342" | "choiceAdventure343" | "choiceAdventure344" | "choiceAdventure345" | "choiceAdventure346" | "choiceAdventure347" | "choiceAdventure348" | "choiceAdventure349" | "choiceAdventure350" | "choiceAdventure351" | "choiceAdventure352" | "choiceAdventure353" | "choiceAdventure354" | "choiceAdventure355" | "choiceAdventure356" | "choiceAdventure357" | "choiceAdventure358" | "choiceAdventure360" | "choiceAdventure361" | "choiceAdventure362" | "choiceAdventure363" | "choiceAdventure364" | "choiceAdventure365" | "choiceAdventure366" | "choiceAdventure376" | "choiceAdventure387" | "choiceAdventure388" | "choiceAdventure389" | "choiceAdventure390" | "choiceAdventure391" | "choiceAdventure392" | "choiceAdventure393" | "choiceAdventure395" | "choiceAdventure396" | "choiceAdventure397" | "choiceAdventure398" | "choiceAdventure399" | "choiceAdventure400" | "choiceAdventure401" | "choiceAdventure402" | "choiceAdventure403" | "choiceAdventure423" | "choiceAdventure424" | "choiceAdventure425" | "choiceAdventure426" | "choiceAdventure427" | "choiceAdventure428" | "choiceAdventure429" | "choiceAdventure430" | "choiceAdventure431" | "choiceAdventure432" | "choiceAdventure433" | "choiceAdventure435" | "choiceAdventure438" | "choiceAdventure439" | "choiceAdventure444" | "choiceAdventure445" | "choiceAdventure446" | "choiceAdventure447" | "choiceAdventure448" | "choiceAdventure449" | "choiceAdventure451" | "choiceAdventure457" | "choiceAdventure458" | "choiceAdventure460" | "choiceAdventure461" | "choiceAdventure462" | "choiceAdventure463" | "choiceAdventure464" | "choiceAdventure465" | "choiceAdventure467" | "choiceAdventure468" | "choiceAdventure469" | "choiceAdventure470" | "choiceAdventure471" | "choiceAdventure472" | "choiceAdventure473" | "choiceAdventure474" | "choiceAdventure475" | "choiceAdventure477" | "choiceAdventure478" | "choiceAdventure480" | "choiceAdventure483" | "choiceAdventure484" | "choiceAdventure485" | "choiceAdventure486" | "choiceAdventure488" | "choiceAdventure489" | "choiceAdventure490" | "choiceAdventure491" | "choiceAdventure496" | "choiceAdventure497" | "choiceAdventure502" | "choiceAdventure503" | "choiceAdventure504" | "choiceAdventure505" | "choiceAdventure506" | "choiceAdventure507" | "choiceAdventure509" | "choiceAdventure510" | "choiceAdventure511" | "choiceAdventure512" | "choiceAdventure513" | "choiceAdventure514" | "choiceAdventure515" | "choiceAdventure517" | "choiceAdventure519" | "choiceAdventure521" | "choiceAdventure522" | "choiceAdventure523" | "choiceAdventure527" | "choiceAdventure528" | "choiceAdventure529" | "choiceAdventure530" | "choiceAdventure531" | "choiceAdventure532" | "choiceAdventure533" | "choiceAdventure534" | "choiceAdventure535" | "choiceAdventure536" | "choiceAdventure538" | "choiceAdventure539" | "choiceAdventure542" | "choiceAdventure543" | "choiceAdventure544" | "choiceAdventure546" | "choiceAdventure548" | "choiceAdventure549" | "choiceAdventure550" | "choiceAdventure551" | "choiceAdventure552" | "choiceAdventure553" | "choiceAdventure554" | "choiceAdventure556" | "choiceAdventure557" | "choiceAdventure558" | "choiceAdventure559" | "choiceAdventure560" | "choiceAdventure561" | "choiceAdventure562" | "choiceAdventure563" | "choiceAdventure564" | "choiceAdventure565" | "choiceAdventure566" | "choiceAdventure567" | "choiceAdventure568" | "choiceAdventure569" | "choiceAdventure571" | "choiceAdventure572" | "choiceAdventure573" | "choiceAdventure574" | "choiceAdventure575" | "choiceAdventure576" | "choiceAdventure577" | "choiceAdventure578" | "choiceAdventure579" | "choiceAdventure581" | "choiceAdventure582" | "choiceAdventure583" | "choiceAdventure584" | "choiceAdventure594" | "choiceAdventure595" | "choiceAdventure596" | "choiceAdventure597" | "choiceAdventure598" | "choiceAdventure599" | "choiceAdventure600" | "choiceAdventure603" | "choiceAdventure604" | "choiceAdventure616" | "choiceAdventure634" | "choiceAdventure654" | "choiceAdventure655" | "choiceAdventure656" | "choiceAdventure657" | "choiceAdventure658" | "choiceAdventure664" | "choiceAdventure669" | "choiceAdventure670" | "choiceAdventure671" | "choiceAdventure672" | "choiceAdventure673" | "choiceAdventure674" | "choiceAdventure675" | "choiceAdventure676" | "choiceAdventure677" | "choiceAdventure678" | "choiceAdventure679" | "choiceAdventure681" | "choiceAdventure683" | "choiceAdventure684" | "choiceAdventure685" | "choiceAdventure686" | "choiceAdventure687" | "choiceAdventure688" | "choiceAdventure689" | "choiceAdventure694" | "choiceAdventure695" | "choiceAdventure696" | "choiceAdventure697" | "choiceAdventure698" | "choiceAdventure706" | "choiceAdventure707" | "choiceAdventure708" | "choiceAdventure709" | "choiceAdventure710" | "choiceAdventure711" | "choiceAdventure712" | "choiceAdventure713" | "choiceAdventure714" | "choiceAdventure715" | "choiceAdventure716" | "choiceAdventure717" | "choiceAdventure721" | "choiceAdventure725" | "choiceAdventure729" | "choiceAdventure733" | "choiceAdventure737" | "choiceAdventure741" | "choiceAdventure745" | "choiceAdventure749" | "choiceAdventure753" | "choiceAdventure771" | "choiceAdventure778" | "choiceAdventure780" | "choiceAdventure781" | "choiceAdventure783" | "choiceAdventure784" | "choiceAdventure785" | "choiceAdventure786" | "choiceAdventure787" | "choiceAdventure788" | "choiceAdventure789" | "choiceAdventure791" | "choiceAdventure793" | "choiceAdventure805" | "choiceAdventure808" | "choiceAdventure809" | "choiceAdventure813" | "choiceAdventure815" | "choiceAdventure851" | "choiceAdventure852" | "choiceAdventure853" | "choiceAdventure854" | "choiceAdventure873" | "choiceAdventure881" | "choiceAdventure882" | "choiceAdventure888" | "choiceAdventure889" | "choiceAdventure921" | "choiceAdventure930" | "choiceAdventure931" | "choiceAdventure932" | "choiceAdventure941" | "choiceAdventure942" | "choiceAdventure943" | "choiceAdventure944" | "choiceAdventure945" | "choiceAdventure946" | "choiceAdventure950" | "choiceAdventure957" | "choiceAdventure958" | "choiceAdventure959" | "choiceAdventure960" | "choiceAdventure961" | "choiceAdventure962" | "choiceAdventure963" | "choiceAdventure964" | "choiceAdventure965" | "choiceAdventure966" | "choiceAdventure970" | "choiceAdventure973" | "choiceAdventure974" | "choiceAdventure975" | "choiceAdventure976" | "choiceAdventure977" | "choiceAdventure979" | "choiceAdventure980" | "choiceAdventure981" | "choiceAdventure982" | "choiceAdventure983" | "choiceAdventure988" | "choiceAdventure989" | "choiceAdventure993" | "choiceAdventure1000" | "choiceAdventure1003" | "choiceAdventure1005" | "choiceAdventure1006" | "choiceAdventure1007" | "choiceAdventure1008" | "choiceAdventure1009" | "choiceAdventure1010" | "choiceAdventure1011" | "choiceAdventure1012" | "choiceAdventure1013" | "choiceAdventure1016" | "choiceAdventure1017" | "choiceAdventure1020" | "choiceAdventure1021" | "choiceAdventure1022" | "choiceAdventure1023" | "choiceAdventure1027" | "choiceAdventure1028" | "choiceAdventure1029" | "choiceAdventure1030" | "choiceAdventure1031" | "choiceAdventure1032" | "choiceAdventure1033" | "choiceAdventure1034" | "choiceAdventure1035" | "choiceAdventure1037" | "choiceAdventure1038" | "choiceAdventure1040" | "choiceAdventure1041" | "choiceAdventure1042" | "choiceAdventure1044" | "choiceAdventure1045" | "choiceAdventure1046" | "choiceAdventure1048" | "choiceAdventure1051" | "choiceAdventure1052" | "choiceAdventure1053" | "choiceAdventure1054" | "choiceAdventure1055" | "choiceAdventure1056" | "choiceAdventure1057" | "choiceAdventure1059" | "choiceAdventure1060" | "choiceAdventure1061" | "choiceAdventure1065" | "choiceAdventure1067" | "choiceAdventure1068" | "choiceAdventure1069" | "choiceAdventure1070" | "choiceAdventure1071" | "choiceAdventure1073" | "choiceAdventure1077" | "choiceAdventure1081" | "choiceAdventure1082" | "choiceAdventure1083" | "choiceAdventure1084" | "choiceAdventure1085" | "choiceAdventure1094" | "choiceAdventure1095" | "choiceAdventure1096" | "choiceAdventure1097" | "choiceAdventure1102" | "choiceAdventure1106" | "choiceAdventure1107" | "choiceAdventure1108" | "choiceAdventure1110" | "choiceAdventure1114" | "choiceAdventure1115" | "choiceAdventure1116" | "choiceAdventure1118" | "choiceAdventure1120" | "choiceAdventure1121" | "choiceAdventure1122" | "choiceAdventure1123" | "choiceAdventure1171" | "choiceAdventure1172" | "choiceAdventure1173" | "choiceAdventure1174" | "choiceAdventure1175" | "choiceAdventure1193" | "choiceAdventure1195" | "choiceAdventure1196" | "choiceAdventure1197" | "choiceAdventure1198" | "choiceAdventure1199" | "choiceAdventure1202" | "choiceAdventure1203" | "choiceAdventure1204" | "choiceAdventure1205" | "choiceAdventure1206" | "choiceAdventure1207" | "choiceAdventure1208" | "choiceAdventure1209" | "choiceAdventure1210" | "choiceAdventure1211" | "choiceAdventure1212" | "choiceAdventure1213" | "choiceAdventure1214" | "choiceAdventure1215" | "choiceAdventure1219" | "choiceAdventure1222" | "choiceAdventure1223" | "choiceAdventure1224" | "choiceAdventure1225" | "choiceAdventure1226" | "choiceAdventure1227" | "choiceAdventure1228" | "choiceAdventure1229" | "choiceAdventure1236" | "choiceAdventure1237" | "choiceAdventure1238" | "choiceAdventure1239" | "choiceAdventure1240" | "choiceAdventure1241" | "choiceAdventure1242" | "choiceAdventure1243" | "choiceAdventure1244" | "choiceAdventure1245" | "choiceAdventure1246" | "choiceAdventure1247" | "choiceAdventure1248" | "choiceAdventure1249" | "choiceAdventure1250" | "choiceAdventure1251" | "choiceAdventure1252" | "choiceAdventure1253" | "choiceAdventure1254" | "choiceAdventure1255" | "choiceAdventure1256" | "choiceAdventure1266" | "choiceAdventure1280" | "choiceAdventure1281" | "choiceAdventure1282" | "choiceAdventure1283" | "choiceAdventure1284" | "choiceAdventure1285" | "choiceAdventure1287" | "choiceAdventure1288" | "choiceAdventure1289" | "choiceAdventure1290" | "choiceAdventure1291" | "choiceAdventure1292" | "choiceAdventure1293" | "choiceAdventure1294" | "choiceAdventure1295" | "choiceAdventure1296" | "choiceAdventure1297" | "choiceAdventure1298" | "choiceAdventure1299" | "choiceAdventure1300" | "choiceAdventure1301" | "choiceAdventure1302" | "choiceAdventure1303" | "choiceAdventure1304" | "choiceAdventure1305" | "choiceAdventure1307" | "choiceAdventure1310" | "choiceAdventure1312" | "choiceAdventure1313" | "choiceAdventure1314" | "choiceAdventure1315" | "choiceAdventure1316" | "choiceAdventure1317" | "choiceAdventure1318" | "choiceAdventure1319" | "choiceAdventure1321" | "choiceAdventure1322" | "choiceAdventure1323" | "choiceAdventure1324" | "choiceAdventure1325" | "choiceAdventure1326" | "choiceAdventure1327" | "choiceAdventure1328" | "choiceAdventure1332" | "choiceAdventure1333" | "choiceAdventure1335" | "choiceAdventure1340" | "choiceAdventure1341" | "choiceAdventure1389" | "choiceAdventure1392" | "choiceAdventure1399" | "choiceAdventure1405" | "choiceAdventure1411" | "choiceAdventure1415";
+export type FamiliarProperty = "commaFamiliar" | "nextQuantumFamiliar" | "preBlackbirdFamiliar";
+export type StatProperty = "nsChallenge1" | "snojoSetting";
