+++ tools/parseDefaultProperties.ts
@@ -1,4 +1,6 @@
 import fetch from "node-fetch";
+import { writeFile } from "fs/promises";
+import path from "path";
 import * as _importForGlobals from "kolmafia";
 
 import {
@@ -13,6 +15,8 @@ import {
 
 const PROPS_FILE =
   "https://sourceforge.net/p/kolmafia/code/HEAD/tree/src/data/defaults.txt?format=raw";
+
+const TYPES_FILE = path.join(__dirname, "../src/propertyTypes.ts")
 
 async function main() {
   const response = await fetch(PROPS_FILE);
@@ -55,11 +59,14 @@ async function main() {
     }
   }
 
+  let contents = `/** THIS FILE IS AUTOMATICALLY GENERATED. See tools/parseDefaultProperties.ts for more information */\n`;
+
   Object.entries(propTypes).forEach(([type, values]) => {
-    console.log(
-      `export type ${type} = ${values.map((v) => `"${v}"`).join(" | ")};`
-    );
+    console.log(`Storing ${values.length} props of type ${type}`);
+    contents += `export type ${type} = ${values.map((v) => `"${v}"`).join(" | ")};\n`;
   });
+
+  await writeFile(TYPES_FILE, contents);
 }
 
 main();
